backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finance-tracker-backend"
version = "0.1.0"
description = "Finance Tracker Family App — Backend API"
requires-python = ">=3.12"

dependencies = [
    # Core framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Schema & config
    "pydantic>=2.7.0",
    "pydantic[email]>=2.7.0",
    "email-validator>=2.1.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "argon2-cffi>=23.1.0",
    "python-multipart>=0.0.9",

    # Logging
    "structlog>=24.1.0",
    
    # Utilities
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
    "aiosqlite>=0.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py312"
